wip refactor test script
